Refactor NavigationComponent logic to its component
